Submission of  Registration, Login and Admin pages
As of now we completed the Login & registration parts and as Dedipya requested we just created navigation side bar in Admin.html so that you guys can add your requirements to that side bar. And moreover we  added particles to Admin page too. And I request to all members not to share the particles code to anyone 🙂 cause we struggle a lot to find and arrange them in our project. for any other queries or changes you can  contact me or @Jesh-17

Co-Authored-By: Jesh-17 <83612190+Jesh-17@users.noreply.github.com>